requirements.txt
pyserial>=3.5
requests>=2.28.0
python-dotenv>=1.0.0 